requirements.txt
streamlit>=1.24.1
numpy>=1.25.0
fpdf2>=2.5.5
